# [FIX] stock, point_of_sale: speedup closing sessions