Speedup closing pos_sessions where there a lots of products tracked by unique serial numbers. This commits speeds up `stock.move.line._apply_putaway_strategy` and `stock.picking._add_mls_related_to_order`.

For _apply_putaway_strategy, the bottleneck without packages comes from the calls `excluded_smls.ids`, with `ids` being a property and `excluded_smls -= sml`. When excluded_smls is large (>10000), doing each one of these calls once by sml stacks up to slowing down the whole process. This commit calls `excluded_smls.ids` once and then use `list.pop()` to remove the current sml's id. This avoids calls to `BaseModel.__sub__`.

For `_add_mls_related_to_order`, this commit removes the call to write on newly created move lines and instead updates ml_vals with qty_fname before creation. This was split in commit bc59c4eb7ef6bc to fix an issue with the "Ship Later" feature. But another fix was introduced in commit d1218aeb4e52deed and the corresponding unittest was removed. Updating ml_vals before creation as it was before the first commit should work correctly. Removing the write avoids calls to modified/recomputes, providing an overall speedup.

#### speedup

Customer 16.0 database with 230k pos.orders, 1M stock.moves, 1M stock.move.lines, 12 000 products.

##### `_apply_putaway_strategy` without result packages timings

| # smls | Before PR | After PR |
|:-------:|:-----------:|:----------:|
| 100 | 94ms | 88ms |
| 1000 | 361ms | 260ms|
| 10000 | 8.37s | 2.27s |
| 100000 | 15min | 16s |
|  151589 | +20min (killed) | 22s |

##### `_add_mls_related_to_order`

In the same database, removing the write calls saves 4min on session closing with 151589 smls.



---
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr