@robodoo 2024-01-15
[Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/149361).